# Regular Error Handling in WKB Parsing