## Description

Previously we used a monad style of error handling. While terse, it was also error prone (e.g. accidentally operating on invalid state). The more regular error handling style is much clearer.

## Check List

Have you:

- Added unit tests? N/A, existing tests.

- Add cmprefimpl tests? (if appropriate?) N/A

## Related Issue

- Fixes https://github.com/peterstace/simplefeatures/issues/247

## Benchmark Results

The numbers are worse, but not statistically significant (these particular benchmarks are very noisy):

```
COMPARISON
name                            old time/op    new time/op    delta
UnmarshalWKB/polygon/n=10-2        466ns ±12%     481ns ±19%   ~     (p=0.369 n=13+15)
UnmarshalWKB/polygon/n=100-2       836ns ±11%     862ns ±22%   ~     (p=0.382 n=14+13)
UnmarshalWKB/polygon/n=1000-2     4.88µs ±14%    4.72µs ±22%   ~     (p=0.264 n=13+14)
UnmarshalWKB/polygon/n=10000-2    44.2µs ± 9%    46.4µs ±15%   ~     (p=0.118 n=12+14)

name                            old alloc/op   new alloc/op   delta
UnmarshalWKB/polygon/n=10-2         284B ± 0%      284B ± 0%   ~     (all equal)
UnmarshalWKB/polygon/n=100-2      1.90kB ± 0%    1.90kB ± 0%   ~     (all equal)
UnmarshalWKB/polygon/n=1000-2     16.5kB ± 0%    16.5kB ± 0%   ~     (all equal)
UnmarshalWKB/polygon/n=10000-2     164kB ± 0%     164kB ± 0%   ~     (all equal)

name                            old allocs/op  new allocs/op  delta
UnmarshalWKB/polygon/n=10-2         7.00 ± 0%      7.00 ± 0%   ~     (all equal)
UnmarshalWKB/polygon/n=100-2        7.00 ± 0%      7.00 ± 0%   ~     (all equal)
UnmarshalWKB/polygon/n=1000-2       7.00 ± 0%      7.00 ± 0%   ~     (all equal)
UnmarshalWKB/polygon/n=10000-2      7.00 ± 0%      7.00 ± 0%   ~     (all equal)
```